clean_up_files moved client
apollo client is now using apollo boost. Kept the deprecated
apollo-client just in case there is a need for it in the near future in
temp
