+++ src/temp/client.js
@@ -1,9 +1,10 @@
+/* APOLLO BOOST HAS REPLACED APOLLO_CLIENT. THIS IS THE OLD WAY */
 /* eslint no-undef: 0 */
 
 import ApolloClient from 'apollo-client';
+import { setContext } from 'apollo-link-context';
 import { createHttpLink } from 'apollo-link-http';
 import { InMemoryCache } from 'apollo-cache-inmemory';
-import { setContext } from 'apollo-link-context';
 
 const token = process.env.REACT_APP_GROUPY_TOKEN;
 
@@ -20,7 +21,6 @@ const httpLink = createHttpLink({
 });
 
 const authLink = setContext((_, { headers }) => {
-  // const token = localStorage.getItem('token');
   return {
     headers: {
       ...headers,
